Конечная

+++ api/views.py
@@ -1,4 +1,3 @@
-from django.shortcuts import render
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from .models import Note
@@ -41,40 +40,58 @@ def getRoutes(request):
     return Response(routes)
 
 
-@api_view(["GET"])
+@api_view(["GET", "POST"])
 def getNotes(request):
-    notes = Note.objects.all()
-    serializer = NoteSerializer(notes, many=True)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        notes = Note.objects.all()
+        serializer = NoteSerializer(notes, many=True)
+        return Response(serializer.data)
+    if request.method == 'POST':
+        data = request.data
+        note = Note.objects.create(body=data['body'])
+        serializer = NoteSerializer(note)
+        return Response(serializer.data)
 
 
-@api_view(["GET"])
+@api_view(["GET", 'PUT', 'DELETE'])
 def getNote(request, pk):
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(note)
-    return Response(serializer.data)
+    if request.method == 'GET':    
+        note = Note.objects.get(id=pk)
+        serializer = NoteSerializer(note)
+        return Response(serializer.data)
+    if request.method == 'PUT':    
+        data = request.data
+        note = Note.objects.get(id=pk)
+        serializer = NoteSerializer(note, data=data)
+        if serializer.is_valid():
+            serializer.save()
+        return Response(serializer.data)
+    if request.method == 'DELETE':    
+        note = Note.objects.get(id=pk)
+        note.delete()
+        return Response('Note was deleted')
 
-@api_view(["POST"])
-def createNote(request):
-    data = request.data
-    note = Note.objects.create(body=data['body'])
-    serializer = NoteSerializer(note)
-    return Response(serializer.data)
+# @api_view(["POST"])
+# def createNote(request):
+#     data = request.data
+#     note = Note.objects.create(body=data['body'])
+#     serializer = NoteSerializer(note)
+#     return Response(serializer.data)
 
 
-@api_view(["PUT"])    
-def updateNote(request, pk):
-    data = request.data
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(note, data=data)
-    if serializer.is_valid():
-        serializer.save()
+# @api_view(["PUT"])    
+# def updateNote(request, pk):
+#     data = request.data
+#     note = Note.objects.get(id=pk)
+#     serializer = NoteSerializer(note, data=data)
+#     if serializer.is_valid():
+#         serializer.save()
 
-    return Response(serializer.data)
+#     return Response(serializer.data)
 
-@api_view(["DELETE"])
-def deleteNote(request, pk):
-    note = Note.objects.get(id=pk)
-    note.delete()
-    return Response('Note was deleted')
+# @api_view(["DELETE"])
+# def deleteNote(request, pk):
+#     note = Note.objects.get(id=pk)
+#     note.delete()
+#     return Response('Note was deleted')
